// src/app/store/parametri/parametri.types.ts
export enum ParametriTypes {
  GET_ALL_VRSTE_REQUEST = '[Parametri] GET_ALL Vrste Request',
  GET_ALL_VRSTE_REQUEST_SUCCESS = '[Parametri]  GET_ALL Vrste Success',
  GET_ALL_VRSTE_REQUEST_FAILURE = '[Login]  GET_ALL Vrste Failure',
  CREATE_VRSTE_REQUEST = '[Parametri] CREATE Vrste Request',
  CREATE_VRSTE_REQUEST_SUCCESS = '[Parametri]  CREATE Vrste Success',
  CREATE_VRSTE_REQUEST_FAILURE = '[Login]  GCREATE Vrste Failure',
  UPDATE_VRSTE_REQUEST = '[Parametri] UPDATE Vrste Request',
  UPDATE_VRSTE_REQUEST_SUCCESS = '[Parametri]  UPDATE Vrste Success',
  UPDATE_VRSTE_REQUEST_FAILURE = '[Login]  UPDATE Vrste Failure',
}